requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0

# DB / Cloud clients
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
google-cloud-bigquery>=3.10.0
db-dtypes>=1.1.1

# HTML processing
beautifulsoup4>=4.12.2

# ML stack
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.5.0  # 모델 튜닝 사용 시 필요
statsmodels>=0.14.0  # 로지스틱 회귀 p-value 산출 시 필요 (없으면 자동 폴백)

# Plotting & Explainability (선택)
matplotlib>=3.7.0  # SHAP 시각화 저장용
shap>=0.44.0       # 선택: summary_plot 사용 시

# API/HTTP
requests>=2.31.0

# OpenAI API
openai>=1.40.0  # openai-python v1 (OpenAI(api_key=...))

# Misc
python-dateutil>=2.8.2
